# fe-pickup/openapi.yaml
openapi: 3.1.0
info:
  title: PickUp API
  description: PickUp API 문서
  version: v1
servers:
- url: /
...
paths:
...
  /auctions:
    get:
      tags:
      - Auction
      summary: 경매 목록 조회
      description: |
        검색어(q), 경매 상태(status), 정렬(sort) 조건으로 경매 목록을 커서 기반으로 조회합니다.
        limit이 있으면 커서/hasNext 없이 상위 N개만 반환하는 홈 노출 전용 모드로 동작합니다.
      operationId: searchAuctions
      parameters:
      - name: memberId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: searchAuctionsRequest
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/SearchAuctionsRequest"
      responses:
        "200":
          description: 경매 목록 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SearchAuctionsPageResponse"
              examples:
                경매 목록 조회 결과:
                  description: 경매 목록 조회 결과
                  value:
                    hasNext: true
                    cursor: string
                    size: 20
                    items:
                    - auctionId: 1
                      consignmentId: 100
                      card:
                        cardId: 10
                        cardName: 리자몽 1st Edition Holo
                        setName: Base Set
                        cardNumber: 4/102
                        language: 일본어
                        rarity: MINT
                        imageUrl: https://example.com/cards/10.png
                      grade: PSA 10
                      auctionStatus: SCHEDULED
                      startingPrice: 10000
                      currentPrice: null
                      startedAt: 2026-08-01T10:00:00
                      endedAt: null
                      remainingSeconds: null
                      watchCount: 0
                      watched: false
                      thumbnailUrl: https://example.com/consignments/100-front.png
        "400":
          description: "요청 값 검증 실패 (잘못된 sort/status/cursor, size/limit 오류 등)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
    post:
      tags:
      - Auction
      summary: 경매 신청
      description: |
        위탁상품에 대한 경매 개최를 신청합니다. 신청이 접수되면 경매 상태는
        SCHEDULED(예정)로 생성되고, 위탁상품 상태는 IN_AUCTION으로 전환됩니다.
        bidIncrement(입찰 단위)는 시작가의 5%로 시스템이 결정합니다.
      operationId: registerAuction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuctionRequest"
            examples:
              경매 신청 요청:
                description: 경매 신청 요청
                value:
                  consignmentId: 100
                  startingPrice: 10000
                  reserve: 15000
                  scheduledStartAt: 2026-08-01T21:00:00
        required: true
      responses:
        "201":
          description: 경매 신청 완료
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateAuctionResponse"
              examples:
                경매 신청 결과:
                  description: 경매 신청 결과
                  value:
                    auctionId: 1
                    consignmentId: 100
                    auctionStatus: SCHEDULED
                    startingPrice: 10000
                    bidIncrement: 500
                    startedAt: 2026-08-01T21:00:00
                    endedAt: 2026-08-08T21:00:00
                    winningBidId: null
                    winningPrice: null
                    createdAt: 2026-07-29T12:00:00
        "400":
          description: "요청 값 검증 실패 (필수 값 누락, 시작가/최소 낙찰가 오류, 과거 일정 등)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "403":
          description: 본인 소유의 위탁상품이 아님
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "404":
          description: 위탁상품을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "409":
          description: 이미 경매 진행/예정 중이거나 신청 불가 상태
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
...
  /auctions/{auctionId}:
    get:
      tags:
      - Auction
      summary: 경매 상세 조회
      description: "경매 ID로 경매 상세 정보(카드, 인증서, 이미지, 판매자, 입찰 관련 정보 등)를 조회합니다."
      operationId: getAuctionDetail
      parameters:
      - name: memberId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: auctionId
        in: path
        description: 경매 ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: 경매 상세 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuctionDetailResponse"
              examples:
                경매 상세 조회 결과:
                  description: 경매 상세 조회 결과
                  value:
                    auctionId: 1
                    consignmentId: 100
                    card:
                      cardId: 10
                      cardName: 리자몽 1st Edition Holo
                      setName: Base Set
                      cardNumber: 4/102
                      language: 일본어
                      rarity: MINT
                      imageUrl: https://example.com/cards/10.png
                    grade: PSA 10
                    auctionStatus: SCHEDULED
                    startingPrice: 10000
                    currentPrice: null
                    startedAt: 2026-08-01T10:00:00
                    endedAt: null
                    remainingSeconds: null
                    watchCount: 0
                    watched: false
                    thumbnailUrl: https://example.com/consignments/100-front.png
                    sellerNickname: 카드마스터샵
                    certificate:
                      certificateId: 1
                      serialNumber: PSA-84213907
                      certificationBody: PSA
                      grade: "10"
                      inspectedAt: 2026-06-30
                    images:
                    - consignmentImageId: 1
                      imageOrder: 0
                      imageUrl: https://example.com/consignments/100-front.png
                    cardState: Gem Mint
                    majorDefect: null
                    bidIncrement: 500
                    nextMinBid: 10000
                    recommendedBid: null
        "404":
          description: 경매를 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
  /auctions/featured:
    get:
      tags:
      - Auction
      summary: 대표 경매 조회
      description: 홈 화면에 노출할 대표 경매 1건을 조회합니다. 진행 중(ONGOING)인 경매 중 관심 수가 가장 많은 경매를
        반환합니다.
      operationId: getFeaturedAuction
      responses:
        "200":
          description: 대표 경매 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuctionListItemResponse"
              examples:
                대표 경매 조회 결과:
                  description: 대표 경매 조회 결과
                  value:
                    auctionId: 1
                    consignmentId: 100
                    card:
                      cardId: 10
                      cardName: 리자몽 1st Edition Holo
                      setName: Base Set
                      cardNumber: 4/102
                      language: 일본어
                      rarity: MINT
                      imageUrl: https://example.com/cards/10.png
                    grade: PSA 10
                    auctionStatus: ONGOING
                    startingPrice: 10000
                    currentPrice: 12000
                    startedAt: 2026-08-01T10:00:00
                    endedAt: 2026-08-01T12:00:00
                    remainingSeconds: 3600
                    watchCount: 42
                    watched: false
                    thumbnailUrl: https://example.com/consignments/100-front.png
        "404":
          description: 대표로 보여줄 진행 중인 경매가 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
components:
  schemas:
...
    ExceptionResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CertificateResponse:
      type: object
      properties:
        certificateId:
          type: integer
          format: int64
        serialNumber:
          type: string
        certificationBody:
          type: string
          enum:
          - PSA
          - BGS
          - CGC
          - SGC
          - ACE
        grade:
          type: string
        inspectedAt:
          type: string
          format: date
...
    CreateAuctionRequest:
      type: object
      properties:
        consignmentId:
          type: integer
          format: int64
        startingPrice:
          type: integer
          format: int64
        reserve:
          type: integer
          format: int64
        scheduledStartAt:
          type: string
          format: date-time
      required:
      - consignmentId
      - reserve
      - scheduledStartAt
      - startingPrice
    CreateAuctionResponse:
      type: object
      properties:
        auctionId:
          type: integer
          format: int64
        consignmentId:
          type: integer
          format: int64
        auctionStatus:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - WON
          - PASSED
        startingPrice:
          type: integer
          format: int64
        bidIncrement:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        winningBidId:
          type: integer
          format: int64
        winningPrice:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
...
    ConsignmentImageResponse:
      type: object
      properties:
        consignmentImageId:
          type: integer
          format: int64
        imageOrder:
          type: integer
          format: int32
        imageUrl:
          type: string
    GetCardDetailResponse:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        cardName:
          type: string
        setName:
          type: string
        cardNumber:
          type: string
        language:
          type: string
        rarity:
          type: string
        imageUrl:
          type: string
...
    SearchAuctionsPageResponse:
      type: object
      properties:
        hasNext:
          type: boolean
        cursor:
          type: string
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuctionListItemResponse"
...
    SearchAuctionsRequest:
      type: object
      properties:
        q:
          type: string
        status:
          type: array
          items:
            type: string
        sort:
          type: string
        limit:
          type: integer
          format: int32
        cursor:
          type: string
        size:
          type: integer
          format: int32
    AuctionDetailResponse:
      type: object
      properties:
        auctionId:
          type: integer
          format: int64
        consignmentId:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/GetCardDetailResponse"
        grade:
          type: string
        auctionStatus:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - WON
          - PASSED
        startingPrice:
          type: integer
          format: int64
        currentPrice:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        remainingSeconds:
          type: integer
          format: int64
        watchCount:
          type: integer
          format: int64
        watched:
          type: boolean
        thumbnailUrl:
          type: string
        sellerNickname:
          type: string
        certificate:
          $ref: "#/components/schemas/CertificateResponse"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ConsignmentImageResponse"
        cardState:
          type: string
        majorDefect:
          type: string
        bidIncrement:
          type: integer
          format: int64
        nextMinBid:
          type: integer
          format: int64
        recommendedBid:
          type: integer
          format: int64
        myBidWon:
          type: boolean
...
    AuctionListItemResponse:
      type: object
      properties:
        auctionId:
          type: integer
          format: int64
        consignmentId:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/GetCardDetailResponse"
        grade:
          type: string
        auctionStatus:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - WON
          - PASSED
        startingPrice:
          type: integer
          format: int64
        currentPrice:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        remainingSeconds:
          type: integer
          format: int64
        watchCount:
          type: integer
          format: int64
        watched:
          type: boolean
        thumbnailUrl:
          type: string
  securitySchemes:
    access-token-cookie:
      type: apiKey
      description: 로그인 시 발급되는 access-token 쿠키
      name: access-token
      in: cookie